+++ src/Game/Text/pt-BR/TraitText.pt-BR.sjson.lua
@@ -222,6 +222,14 @@ local newData = {
 	-- 	"You gain {#UpgradeFormat}{$TooltipData.StatDisplay1}{#Prev}{!Icons.ReRollAlt}, and can alter {#BoldFormat}Location Rewards{#Prev}. This does not break {$Keywords.Random}.",
 	-- },
 	-- #endregion
+
+	-- #region WellShops
+	{
+		Id = "ModsNikkelMHadesBiomes_TemporaryBlockExplodingChariotsTrait",
+		DisplayName = "Soltura das Rodas Flamejantes",
+		Description = "Previne inimigos {$TooltipData.BlockedEnemyTypes[1]} de aparecerem em {$Keywords.EncounterPlural}.",
+	},
+	-- #endregion
 }
 
 sjson.hook(traitTextFile, function(data)