+++ src/Game/Text/tr/TraitText.tr.sjson.lua
@@ -222,6 +222,14 @@ local newData = {
 	-- 	"You gain {#UpgradeFormat}{$TooltipData.StatDisplay1}{#Prev}{!Icons.ReRollAlt}, and can alter {#BoldFormat}Location Rewards{#Prev}. This does not break {$Keywords.Random}.",
 	-- },
 	-- #endregion
+
+	-- #region WellShops
+	-- {
+	-- 	Id = "ModsNikkelMHadesBiomes_TemporaryBlockExplodingChariotsTrait",
+	-- 	DisplayName = "Flame Wheels Release",
+	-- 	Description = "Prevent {$TooltipData.BlockedEnemyTypes[1]} foes from appearing in {$Keywords.EncounterPlural}.",
+	-- },
+	-- #endregion
 }
 
 sjson.hook(traitTextFile, function(data)