+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v0.7.2
 
 <!--Releasenotes start-->
+- Added "Flame Wheels Release" item to Wells of Charon.
 - After freeing Orpheus, you are guaranteed to encounter him before Sisyphus can appear again.
 - Fixed: When a Skull-Crusher is polymorphed while taking off, its animations break (Skull-Crushers can no longer be polymorphed).
 - Fixed: The button prompt to reroll rewards in Asphodel does not change if the keybinding is changed.